HUD: Use element map in hud receiver to apply skin textures
This removes the need for the body links to have the region
(upper, lower, etc) in the link object description.  This is now
completely handled in the element map ready from the !R2 CONFIG notecard.

Rename ru2_hud_receiver.lsl to r2_hud-receiver.lsl.  We should be able
to share this between both projects with the addition of the
config notecard.